requirements.txt
streamlit
pandas
google-generativeai
python-docx
openpyxl